Move Veryl module discovery into frontend
